claims added

+++ resources/item.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, reqparse
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 
 from models.item import ItemModel
 
@@ -16,7 +16,12 @@ class Item(Resource):
             return {'item': item.json()}, 200
         return {'msg': f'Item does not exist'}, 404
 
+    @jwt_required()
     def post(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'msg':'Missing admin privilage'}, 401
+
         item = ItemModel.get_item_by_name(name)       
         data = Item.parser.parse_args()
         
@@ -32,13 +37,23 @@ class Item(Resource):
 
         return {'item': new_item.json()}, 201
 
+    @jwt_required()
     def delete(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'msg':'Missing admin privilage'}, 401
+
         item = ItemModel.get_item_by_name(name)
         if item:
             item.delete_from_db() 
         return {'msg':'item has been deleted'}
 
+    @jwt_required()
     def put(self, name):
+        claims = get_jwt()
+        if not claims['is_admin']:
+            return {'msg':'Missing admin privilage'}, 401
+            
         item = ItemModel.get_item_by_name(name)
         data = Item.parser.parse_args()
